requirements.txt
discord.py[voice]
git+https://github.com/Rapptz/discord-ext-menus
topggpy
colorlog
lavalink
colorama
pymongo
dnspython